+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - tip: support using custom "text" keybinding to insert any text.
 - tip: enable chewing.auto_snapshot_selections by default.
 - tip: support zh-CN language profile.
+- tip: draw preedit text segments with different underline style.
 - prefs: preferences now only show fonts that can render Chinese characters.
 - prefs: new advanced option to lock Chinese or English mode on CapsLock.
 - prefs: new advanced option to customize candidate list border color.